backend/requirements.txt
django
djangorestframework
django-cors-headers
requests
python-decouple
gunicorn
whitenoise
